+++ test/path/path.test.ts
@@ -1,7 +1,7 @@
-import { Elysia, t } from '../src'
+import { Elysia, t } from '../../src'
 
 import { describe, expect, it } from 'bun:test'
-import { post, req } from './utils'
+import { post, req } from '../utils'
 
 describe('Path', () => {
 	it('handle root', async () => {